csv出力に会計年度、証券コードを加える

+++ ResearchXBRL.Application/DTO/ReverseLookupAccountItems/FinancialReport.cs
@@ -5,7 +5,7 @@ namespace ResearchXBRL.Application.DTO.ReverseLookupAccountItems;
 
 public record FinancialReport
 {
-    public decimal SecuritiesCode { get; init; }
+    public int SecuritiesCode { get; init; }
     public AccountingStandards AccountingStandard { get; init; }
     public DateOnly FiscalYear { get; init; }
     public IReadOnlyDictionary<string, decimal?> AccountAmounts { get; init; } = new Dictionary<string, decimal?>();
